use new CLI approach
